docs: Switch to a SQLite sidecar
Based on the following note:
Can we make SyndDB 100% a normal Rust backend?
As long as it’s deterministic and re-executable, people can just write regular Rust, stick it inside a Docker container, and stick it inside a TEE. We could have special tables that represent message passing, and automatically pick up on those for Bridge.sol

All we then need is a listener that is attached that hooks into the SQLite database and logs/snapshots/monitors it

We don’t need to re-derive oracle data for example, we just need to re-derive the SQL statements

Don’t overthink full verifiability. Just have a listener that attaches to the SQLite DB and watches for changes. The sequencer is already running trusted business logic so full re-derivation of the business logic is not necessary, just re-derivation of the SQL statements

Overall plan:
SQL as verifiability. The sequencer just writes all data to SQLite that is necessary for people to check its results. We don’t attempt to offer full reproducibility from APIs (that is not always possible) but instead have the application write everything necessary for validators to check, which can add their own business logic to the SQL logs before signing off. This makes validators not 1:1 with sequencers, but instead a check on the sequencer’s DB.

We can even accomplish this by having a reader attach to the SQLite DB and automatically monitor for new logs that it posts to the DA layer. No customization will be needed, the reader can be run as a sidecar that connects to the DB.